+++ apps/api/src/app/endpoints/public/public.controller.ts
@@ -57,7 +57,7 @@ export class PublicController {
     }
 
     const [
-      { holdings, markets },
+      { createdAt, holdings, markets },
       { performance: performance1d },
       { performance: performanceMax },
       { performance: performanceYtd }
@@ -81,6 +81,7 @@ export class PublicController {
     });
 
     const publicPortfolioResponse: PublicPortfolioResponse = {
+      createdAt,
       hasDetails,
       markets,
       alias: access.alias,