+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator.ts
@@ -104,6 +104,7 @@ export class TWRPortfolioCalculator extends PortfolioCalculator {
       activitiesCount: this.activities.filter(({ type }) => {
         return ['BUY', 'SELL'].includes(type);
       }).length,
+      createdAt: new Date(),
       errors: [],
       historicalData: [],
       totalLiabilitiesWithCurrencyEffect: new Big(0),