+++ libs/common/src/lib/models/portfolio-snapshot.ts
@@ -11,6 +11,8 @@ import { Transform, Type } from 'class-transformer';
 export class PortfolioSnapshot {
   activitiesCount: number;
 
+  createdAt: Date;
+
   @Transform(transformToBig, { toClassOnly: true })
   @Type(() => Big)
   currentValueInBaseCurrency: Big;