+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Extended the export functionality by the platforms
+- Extended the portfolio snapshot in the portfolio calculator by the `createdAt` timestamp
 - Extended the _Trackinsight_ data enhancer for asset profile data by `cusip`
 - Added _Storybook_ to the build process